+++ judge/contest_format/__init__.py
@@ -3,5 +3,5 @@ from judge.contest_format.default import DefaultContestFormat
 from judge.contest_format.ecoo import ECOOContestFormat
 from judge.contest_format.icpc import ICPCContestFormat
 from judge.contest_format.ioi import IOIContestFormat
-from judge.contest_format.ioi16 import IOI16ContestFormat
+from judge.contest_format.legacy_ioi import LegacyIOIContestFormat
 from judge.contest_format.registry import choices, formats
